+++ lib/src/get_state_manager/src/rx_flutter/rx_notifier.dart
@@ -5,7 +5,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter/scheduler.dart';
 
 import '../../../../get_x_master.dart';
-import '../simple/list_notifier.dart';
+
 
 mixin StateMixin<T> on ListNotifierMixin {
   T? _value;